# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/issues/{issue_id}/vote": {
      "post": {
        "tags": [
          "issue-votes"
        ],
        "operationId": "vote_issue",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Vote added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "issue-votes"
        ],
        "operationId": "unvote_issue",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vote removed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/issues/{issue_id}/votes": {
      "get": {
        "tags": [
          "issue-votes"
        ],
        "operationId": "list_votes",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Votes listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/issues/{issue_id}/watch": {
      "post": {
        "tags": [
          "issue-watchers"
        ],
        "operationId": "watch_issue",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Now watching the issue"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Issue not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "issue-watchers"
        ],
        "operationId": "unwatch_issue",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Stopped watching the issue"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/issues/{issue_id}/watchers": {
      "get": {
        "tags": [
          "issue-watchers"
        ],
        "operationId": "list_watchers",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "description": "Issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watchers listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatcherListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoteListResponse": {
        "type": "object",
        "required": [
          "votes",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "votes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteResponse"
            }
          }
        }
      },
      "VoteResponse": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "display_name",
          "voted_at"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "voted_at": {
            "type": "string"
          }
        }
      },
...
      "WatchRequest": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WatcherListResponse": {
        "type": "object",
        "required": [
          "watchers"
        ],
        "properties": {
          "watchers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatcherResponse"
            }
          }
        }
      },
      "WatcherResponse": {
        "type": "object",
        "required": [
          "user_id",
          "username",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  }
}